backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
pandas==2.1.3
reportlab==4.0.7
matplotlib==3.8.2
Pillow==10.1.0
openpyxl==3.1.2
numpy==1.26.2
plotly==5.18.0
drf-spectacular==0.27.0
requests==2.31.0
